+++ src/Ocelot/Claims/Middleware/ClaimsBuilderMiddlewareExtensions.cs
@@ -3,11 +3,11 @@ using Ocelot.Middleware.Pipeline;
 
 namespace Ocelot.Claims.Middleware
 {
-    public static class ClaimsBuilderMiddlewareExtensions
+    public static class ClaimsToClaimsMiddlewareExtensions
     {
-        public static IOcelotPipelineBuilder UseClaimsBuilderMiddleware(this IOcelotPipelineBuilder builder)
+        public static IOcelotPipelineBuilder UseClaimsToClaimsMiddleware(this IOcelotPipelineBuilder builder)
         {
-            return builder.UseMiddleware<ClaimsBuilderMiddleware>();
+            return builder.UseMiddleware<ClaimsToClaimsMiddleware>();
         }
     }
 }
