+++ src/Ocelot/Headers/HttpContextRequestHeaderReplacer.cs
